[fix] 답변 기다리는 질문 페이지 코드 수정
