+++ src/components/question/types.tsx
@@ -9,3 +9,14 @@ export interface ModalProps {
   content: string
   clickModal: () => void
 }
+
+// 질문 props
+export interface QuestionProps {
+  questionId: number
+  content: string
+  nickname: string
+  profileOnOff: boolean
+  createdDate: string
+  isAnswered: boolean
+  fcmtoken: string
+}
