--- src/pages/question/types.ts
@@ -1,9 +1,0 @@
-export interface QuestionProps {
-  questionId: number
-  content: string
-  nickname: string
-  profileOnOff: boolean
-  createdDate: string
-  isAnswered: boolean
-  fcmtoken: string
-}
